backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
requests==2.31.0